fixed use-form

+++ app/routes/admin.roles.$roleId.tsx
@@ -3,11 +3,17 @@ import { json, LoaderFunctionArgs, redirect } from "@remix-run/node";
 import { useLoaderData } from "@remix-run/react";
 import { z } from "zod";
 import { AddOrUpdateRoleForm } from "~/components/admin/add-or-update-role-form";
+import { PermissionsEnum } from "~/lib/config.server";
 import { prisma } from "~/lib/db.server";
 import { roleSchema } from "~/lib/schema";
 import { ca, safeAction } from "~/lib/utils";
 import { getAllPermissions } from "~/models/permission.server";
 import { deleteRole, getRoleById } from "~/models/role.server";
+import { requireStaff, requireStaffId } from "~/session.server";
+import { requirePermissions } from "~/use-cases/permission.server";
+import { ErrorBoundary } from "~/components/shared/error-boudary";
+
+export { ErrorBoundary };
 export const loader = async ({ request, params }: LoaderFunctionArgs) => {
   const id = params.roleId as string;
   const role = await getRoleById(prisma, id);
@@ -22,17 +28,21 @@ export const loader = async ({ request, params }: LoaderFunctionArgs) => {
 export const action = safeAction([
   {
     method: "DELETE",
-    action: ca(async ({ request, params }) => {
+    action: async ({ request, params }) => {
+      const userId = await requireStaffId(request);
+      await requirePermissions(prisma, userId, [PermissionsEnum.DeleteRoles]);
       const id = params.roleId as string;
       if (!id) return json({ success: false, error: "missing id." }, 400);
       await deleteRole(id);
       return json({ success: true });
-    }),
+    },
   },
   {
     method: "PUT",
     schema: roleSchema,
-    action: ca(async ({ request, params }, data) => {
+    action: async ({ request, params }, data) => {
+      const userId = await requireStaffId(request);
+      await requirePermissions(prisma, userId, [PermissionsEnum.UpdateRoles]);
       const id = params.roleId as string;
       if (!id) return json({ success: false, error: "missing id." }, 400);
       await prisma.$transaction(async (db) => {
@@ -67,7 +77,7 @@ export const action = safeAction([
         }
       });
       return json({ success: true });
-    }),
+    },
   },
 ]);
 export default function UpdateRolePage() {
@@ -76,16 +86,16 @@ export default function UpdateRolePage() {
     <>
       <div className="flex justify-between items-center mb-4 sticky top-4 bg-white ">
         <div>
-          <h1 className="text-2xl font-bold">Sửa quyền</h1>
+          <h1 className="text-2xl font-bold">Sửa vai trò</h1>
           <nav className="text-sm text-gray-600">
             <a href="/admin" className="hover:underline">
               Trang chủ
             </a>{" "}
             &gt;{" "}
             <a href="/admin/roles" className="hover:underline">
-              Quản lí quyền
+              Quản lí vai trò
             </a>{" "}
-            &gt; Sửa quyền
+            &gt; Sửa vai trò
           </nav>
         </div>
       </div>
